INTYGFV-15713: Fixed checkstyle
