+++ web/src/main/java/se/inera/intyg/rehabstod/service/communication/UnansweredCommunicationFilterServiceImpl.java
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package se.inera.intyg.rehabstod.service.unansweredCommunication;
+package se.inera.intyg.rehabstod.service.communication;
 
 import org.springframework.stereotype.Service;
 import se.inera.intyg.rehabstod.service.sjukfall.dto.UnansweredCommunicationFilterType;
